requirements.txt
pytorch-lightning==1.2.10
datasets==1.6.2
transformers==4.5.1
scikit-learn==0.24.2
pytest
tensorboard
wandb
torchmetrics
hydra-core 
hydra_colorlog
dvc
'dvc[gdrive]'
'dvc[s3]'
onnxruntime
uvicorn
fastapi